requirements.txt
rank_bm25>=0.2.2
faiss-cpu>=1.7.4
transformers>=4.41.0
sentence-transformers>=3.2
torch>=2.2
scipy>=1.12
scikit-learn>=1.4
pandas>=2.2
matplotlib>=3.9
tqdm>=4.66
pytest>=8
